Rename file names for human player module.
